% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC.R
\name{ABC_on_hyperparameters}
\alias{ABC_on_hyperparameters}
\title{Conduct an ABC on a range of effect size distribution hyperparameters}
\usage{
ABC_on_hyperparameters(
  x,
  phenotypes,
  iters,
  pi_func = function(x) rbeta(x, 25, 1),
  df_func = NULL,
  scale_func = NULL,
  h = NULL,
  julia.path = "julia",
  chain_length = 1e+05,
  burnin = 5000,
  thin = 100,
  method = "BayesB",
  ABC_scheme = "A",
  par = F,
  run_number = NULL,
  est_h = F,
  save_effects = T,
  joint_res = NULL,
  joint_acceptance = NULL,
  joint_res_dist = "ks.D",
  delta = 0.5,
  pcut = 5e-04
)
}
\arguments{
\item{x}{matrix. Input genotypes, SNPs as rows, columns as individuals. Genotypes formatted as 0,1,2 for the major homozygote, heterozygote, and minor homozygote, respectively.}

\item{phenotypes}{numeric vector. Observed phenotypes, one per individual.}

\item{iters}{numeric. Number of ABC permutations to run.}

\item{pi_func}{function, default function(x) rbeta(x, 25, 1). A distribution function for generating pi prior. Should take only one argument (n, the number of iters).}

\item{df_func}{function, default NULL. A distribution function for generating df prior. Should take only one argument (n, the number of iters).numeric vector of length 2, default NULL. Range (min, max) of degrees of freedom values to run.}

\item{scale_func}{function, default NULL. A distribution function for generating scale prior. Should take only one argument (n, the number of iters).numeric vector of length 2, default NULL. Range (min, max) of scale values to run.}

\item{h}{numeric, default NULL. Heritability to use. Will take a range in the future.}

\item{julia.path}{character, defualt "julia". File path to the julia executable, required for JWAS.}

\item{chain_length}{numeric, default 100000. Length of the MCMC chains used in each step of the ABC.}

\item{burnin}{numeric, default 5000. Number of MCMC chain steps discarded at the start of each MCMC.}

\item{thin}{numeric, default 100. Number of MCMC chain steps discarded between each sample used to form the posterior.}

\item{method}{character, default "bayesB". The marker effect size distribution/prediction method to use.}

\item{ABC_scheme}{character, default "A". The ABC_scheme to use. See details.}

\item{par}{numeric or FALSE, default FALSE. If numeric, the number of cores on which to run the ABC.}

\item{run_number}{numeric, default NULL. Controls how the itermediate output directories are named. If numeric, will be named for the number, otherwise, will be named for the iteration.}
}
\description{
Runs Approximate Bayesian Computation across a range of marker effect size
distribution hyperparameters using one of three different schemes in order to determine
the hyperparamters that generate a distribution most like the real genomic architecture of the trait.
}
\details{
ABC schemes: \itemize{
    \item{"A": }{Genomic Data -> prediction with prior hyperparameters -> compare predicted phenotypes to real phenotypes.}
    \item{"B": }{Genomic Data -> generate psuedo marker effects using distribution with prior hyperparameters -> prediction with defaults -> compare predicted phenotypes to real phenotypes.}
    \item{"C": }{Part 1: Genomic Data -> generate psuedo marker effects using distribution with prior hyperparameters -> "pseudo" predicted marker effects.
                 Part 2: Genomic Data -> prediction with defaults -> direct estimated marker effects.
                 Part 3: Compare direct to "pseudo" estimated marker effects.}
}
}
