% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs.R
\name{BL_survival}
\alias{BL_survival}
\title{Calculate survival using the Burger and Lynch method}
\usage{
BL_survival(phenotypes, opt_pheno, omega, ...)
}
\arguments{
\item{phenotypes}{numeric. Vector of phenotypes of individuals.}

\item{opt_pheno}{numeric. Optimum phenotype this generation.}

\item{omega}{numeric. Strength of stabilizing selection. Smaller numbers mean stronger selection.}
}
\description{
Surivial is caluclated using equation 1 from Burger and Lynch 1995.
}
