% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation.R
\name{calc_distribution_stats}
\alias{calc_distribution_stats}
\title{Calculates statistics for GWAS or prediction effect/p-value distributions.}
\usage{
calc_distribution_stats(
  x = NULL,
  meta,
  phenos,
  center = T,
  scheme = "gwas",
  chr = "chr",
  peak_delta = 0.5,
  peak_pcut = 5e-04,
  window_sigma = 50,
  burnin = NULL,
  thin = NULL,
  chain_length = NULL,
  maf = 0.05,
  phased = FALSE,
  pass_windows = F,
  pass_G = NULL,
  GMMAT_infile = NULL,
  find_similar_effects = FALSE,
  real_effects = NULL,
  save_effects = FALSE
)
}
\arguments{
\item{meta}{data.frame. A data.frame containing SNP metadata. Must have a column containing chromosome info (named to match the "chr" argument) and another named "position" containing position information in base pairs.}

\item{phenos}{numeric. A vector of phenotypes.}

\item{scheme}{"gwas" or "gp", default "gwas". Method to use.}

\item{chr}{character, default "chr". Name of the metadata column containing chromosome information.}

\item{find_similar_effects}{logical, default F. If F, calculates descriptive stats for the effects. If T, compares distribution to the "real_effects" data.}

\item{real_effects}{numeric or NULL, default NULL. If provided and find_similar_effects is TRUE, this effect distribution is compared to that for the provided geno/phenotypes. This can
be calculated for the real data using \code{\link{pred_GWAS_FBM}}.}

\item{save_effects}{character or FALSE, default FALSE. If true, simulated effects will be appended to filepath provided here.}

\item{Object}{coercable to matrix or NULL, default NULL. If provided, a matrix coercable object containing genotype calls. If NULL, a GMMAT infile must instead be provided.}
}
\description{
Given genotypic data, calculates descriptive statistics of GWAS/gp results \emph{or} compares
those results to that from the real data.
}
