% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{clean_phenotypes}
\alias{clean_phenotypes}
\title{Remove genotypes without phenotypic data.}
\usage{
clean_phenotypes(genotypes, phenos)
}
\arguments{
\item{genotypes}{numeric matrix, data.table, or data.frame. Input genotypes, two columns per individual.}

\item{phenos}{numeric vector. Phenotypes for each individual. Missing data should be marked with NA.}
}
\description{
Remove genotypes without phenotypic data.
}
