% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_description.R
\name{compare_distributions}
\alias{compare_distributions}
\title{compare two different effect size/phenotype/etc distributions}
\usage{
compare_distributions(o, p)
}
\description{
compare two different effect size/phenotype/etc distributions
}
