% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{effect_size_distributions}
\alias{effect_size_distributions}
\alias{rbayesB}
\alias{rbayesB_fixed}
\alias{rzero_inflated_normal}
\alias{rzero_inflated_normal_fixed}
\alias{rflat}
\alias{rflat_fixed}
\title{Effect size distributions.}
\usage{
rbayesB(n, pi, d.f, scale)

rbayesB_fixed(n, sites, d.f, scale)

rzero_inflated_normal(n, pi, sd, scale)

rzero_inflated_normal_fixed(n, sites, sd, scale)

rflat(n, pi, scale)

rflat_fixed(n, sites, scale)
}
\arguments{
\item{n}{numeric. Number of draws to make from the distribution}

\item{pi}{numeric. Probability that any one site has zero effect}

\item{d.f}{numeric. Degrees of freedom for the effect size distribution.}

\item{scale}{numeric. Scale/shape parameter for the scaled-inverse
chi-squared distribution.}

\item{sites}{numeric. Number of sites that have an effect.}

\item{sd}{numeric. Standard deviation for the effect size distribution.}
}
\description{
Generate any number of random values drawn from the locus effect size
distributions.
}
\section{Functions}{
\itemize{
\item \code{rbayesB}: assign some loci an effect from a scaled t distribution given a probablity of non-zero effects

\item \code{rbayesB_fixed}: assign a fixed number of loci effects from a scaled t distribution given

\item \code{rzero_inflated_normal}: assign some loci an effect from a scaled normal distribution given a probability of non-zero effects

\item \code{rzero_inflated_normal_fixed}: assign a fixed number of loci effects from a scaled normal distribution

\item \code{rflat}: asign loci a fixed effect given a probability of non-zero effects

\item \code{rflat_fixed}: assign a fixed number of loci a fixed effect
}}

\section{BayesB}{

Under a bayesB model, the effect of each site is drawn from a distribution
where var(g) == 0. Each locus has probability \emph{pi} of having no effect. Effects are otherwise drawn from a scaled t
distribution with degrees of freedom \emph{d.f} and scale \emph{scale}.
Alternatively, a fixed number of sites (\emph{sites}) can be given an effect from the same
distribution. This is equivalent to the expected number of sites when \emph{pi} = 1
- (\emph{sites}/\emph{n}), where \emph{n} is the number of loci.
}

\section{scaled normal}{

Effects are drawn from a normal distribution where var(g) == 0.
Each locus has probability \emph{pi} of having no effect. Effects are otherwise drawn from a normal
distribution with mean 0, standard deviation \emph{sd}, and scale \emph{scale}.
Alternatively, a fixed number of sites (\emph{sites}) can be given an effect from the same
distribution. This is equivalent to the expected number of sites when \emph{pi} = 1
- (\emph{sites}/\emph{n}), where \emph{n} is the number of loci.
}

\section{flat}{

Each locus has probability \emph{pi} of having no effect. Effects are otherwise equal to
the parameter scale \emph{scale}.
Alternatively, a fixed number of sites (\emph{sites}) can be given an effect from the same
distribution. This is equivalent to the expected number of sites when \emph{pi} = 1
- (\emph{sites}/\emph{n}), where \emph{n} is the number of loci.
}

