% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation.R
\name{estimate_gen_arch_from_sims}
\alias{estimate_gen_arch_from_sims}
\title{Estimate pi from simulations.}
\usage{
estimate_gen_arch_from_sims(
  x,
  meta,
  phenos,
  sims,
  ABC_res,
  p = 0.25,
  num.trees = 1000,
  mtry = ncol(sims) - 1,
  num.threads = NULL,
  pi_transform = function(pi) log10(1 - pi),
  pi_back_transform = function(pi) 1 - 10^(pi),
  importance = "none",
  scheme = "gwas",
  chr = "chr",
  peak_delta = 0.5,
  peak_pcut = 5e-04,
  window_sigma = 50,
  burnin = NULL,
  thin = NULL,
  chain_length = NULL,
  ABC_acceptance_threshold = 0.005,
  ABC_dist_var = "ks",
  ABC_scale_transform = function(scale) log10(scale),
  ABC_scale_back_transform = function(scale) 10^scale,
  ...
)
}
\arguments{
\item{x}{numeric. Descriptive statistics for the estimated effect sizes/associaion p-values
from the observed data.}

\item{sims}{data.frame. Data.frame that matches that produced by \code{\link{sim_gen}}, containing descriptive statistics for the estimated effect sizes/associaion p-values
from the simulted data as well as a column named 'pi' containing the simulated pi values. Each row should be a single
simulation. May contain other columns titled 'df', 'scale', and 'h'.}

\item{p}{numeric < 1 and > 0, default .25. Proportion of sims to hold out from model estimation for use in cross-evalutation.}

\item{num.trees}{numeric, default 1000. Number of trees to grow during the random forest.}

\item{mtry}{numeric, default ncol(sims) - 1. Number of variables to possibly split at each node during random forest. See
\code{\link[ranger]{ranger}} for details.}

\item{num.threads}{numeric, default NULL. Number of processing threads to use for tree growth and cross-evaluation.}

\item{pi_transform}{function or FALSE, default function(pi) log10(1 - pi). Transformation to use on pi. Usefull if pi values in simulations
are heavily skewed, as is likely given that it is a non-linear parameter.}

\item{importance}{character, default "none". Determines how variable importance is computed, if it is at all. See \code{\link[ranger]{ranger}}
for details. Note that "permutation", while accurate, can be very slow. By and large, this isn't needed for genetic architecture prediciton.}

\item{...}{extra arguments passed to \code{\link[ranger]{ranger}}.}
}
\description{
Estimate pi from distribution descriptions from simulated effect size distributions
using a random forest.
}
\author{
William Hemstrom
}
