% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation.R
\name{estimate_scale_from_ABC}
\alias{estimate_scale_from_ABC}
\title{Estimate scale ABC results}
\usage{
estimate_scale_from_ABC(
  x,
  pi,
  pi_transform = function(pi) log10(1 - pi),
  threshold = 0.005,
  dist_var = "ks",
  quantiles = seq(0 + 1e-04, 1 - 1e-04, by = 1e-04),
  scale_transform = function(scale) log10(scale),
  scale_back_transform = function(scale) 10^scale
)
}
\arguments{
\item{x}{data.frame. ABC results, such as those given by \code{\link{ABC_on_hyperparameters}}.}

\item{pi}{numeric. Vector of pi values for which to calculate scale.}

\item{pi_transform}{function or FALSE, default function(pi) log10(1 - pi). Transformation to use on pi. Usefull if pi values in simulations
are heavily skewed, as is likely given that it is a non-linear parameter.}

\item{threshold}{numeric, default 0.005. Proportion of accepted runs.}

\item{dist_var}{character, default "ks". Name of the distance varaible to use in picking accepted runs.}
}
\description{
Esitmates scale based on ABC results using gam smoothing.
}
