% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC.R
\name{gen_parms}
\alias{gen_parms}
\title{Draw parameter values from the posteriour distribution of an ABC using a density kernal}
\usage{
gen_parms(x, res, num_accepted, parameters, grid = 10000, dist.var = "ks.D")
}
\arguments{
\item{x}{numeric, the number of values to draw.}

\item{res}{data.frame, the results of an ABC run, must include a "dist" column.}

\item{num_accepted}{numeric, either the proportion of runs to accept or the number of runs to accept.}

\item{parameters}{character, the parameter names for which to draw values.}
}
\description{
Draw parameter values from the posteriour distribution of an ABC using a density kernal
}
