% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_description.R
\name{generate_pseudo_effects}
\alias{generate_pseudo_effects}
\title{generate psuedo effects using passed parameters, an effect size distribution, and heritability.}
\usage{
generate_pseudo_effects(x, effect_distribution, parameters, h, center = T)
}
\description{
generate psuedo effects using passed parameters, an effect size distribution, and heritability.
}
