% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs.R
\name{gs}
\alias{gs}
\title{Simulate selection and population growth}
\usage{
gs(
  genotypes,
  meta,
  phenotypes = NULL,
  BVs = NULL,
  h,
  gens,
  chr.length,
  growth.function = function(n) logistic_growth(n, 500, 2),
  survival.function = function(phenotypes, opt_pheno, ...) BL_survival(phenotypes,
    opt_pheno, omega = 1),
  selection.shift.function = function(opt, iv) optimum_constant_slide(opt, iv, 0.3),
  rec.dist = function(n) rpois(n, lambda = 1),
  var.theta = 0,
  plot_during_progress = FALSE,
  facet = "group",
  do.sexes = TRUE,
  init = F,
  verbose = T,
  print.all.freqs = F,
  model = NULL,
  K_thin_post_surv = NULL
)
}
\arguments{
\item{genotypes}{unphased matrix of genotypes, one row per snp and one column
per chromosome/gene copy. Copies from individuals must be sequential.}

\item{meta}{data.frame of snp metadata, with the first column containing
chromosome info, the second containing position in base pairs, and the
third, named "effects" optionally containing effect sizes.}

\item{chr.length}{numeric vector of chromosome lengths \emph{in the same
order as would be given by unique(meta[,1])}.}
}
\description{
Simulate selection and population growth
}
