% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs.R
\name{gs_breeders}
\alias{gs_breeders}
\title{Simulate population demographics under selection with the Breeder's Equation.}
\usage{
gs_breeders(
  phenotypes,
  h,
  B,
  K,
  omega,
  var.theta,
  k,
  gens = Inf,
  n = NULL,
  nloci = NULL,
  alpha = NULL,
  mu = NULL
)
}
\arguments{
\item{phenotypes}{numeric. Vector of initial phenotypes.}

\item{h}{numeric. Heritability of the trait.}

\item{B}{numeric. Number of offspring each surviving adult has. Not stochastic.}

\item{K}{numeric. Carrying capacity. Individuals surviving selection will be randomly culled to this number.}

\item{omega}{numeric. Width of the selection function. Smaller numbers equate to stronger selection around the optimum.}

\item{var.theta}{numeric. Variance of the stochastic selection optimum.}

\item{k}{numeric. Amount by which the optimum phenotype increases each generation (where mean(opt) = k*t).}

\item{gens}{numeric, default Inf. Number of generations to iterate.}

\item{n}{numeric, default NULL. Initial population size. If null, equal to the number of provided phenotypes.}

\item{nloci}{numeric, default NULL. Number of effective loci. If nloci, alpha, and mu are all provided, the stochastic house of cards approximation of genetic variance is used to approximate changes in genetic variance each generation based \emph{only} on changes in effective population size.}

\item{alpha}{numeric, default NULL. Standard deviation of the effect of new mutations. If nloci, alpha, and mu are all provided, the stochastic house of cards approximation of genetic variance is used to approximate changes in genetic variance each generation based \emph{only} on changes in effective population size.}

\item{mu}{numeric, default NULL. Mutation rate. If nloci, alpha, and mu are all provided, the stochastic house of cards approximation of genetic variance is used to approximate changes in genetic variance each generation based \emph{only} on changes in effective population size.}
}
\description{
Uses several equations from Burger and Lynch 1995 to calculate ne and additive genetic variance,
but otherwise simulates purely based on the Breeder's Equation.
}
\details{
If nloci, alpha, and mu are all provided, the stochastic house of cards
approximation of genetic variance is used to approximate changes in genetic variance
each generation based \emph{only} on changes in effective population size. Otherwise,
genetic variance is assumed to not change in each generation and is based solely on the
genetic variance calculated from the provided heritability and phenotypes.
}
