% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation.R
\name{hyperparameter_regression_on_ABC}
\alias{hyperparameter_regression_on_ABC}
\title{Estimate a hyperparamter from ABC results via regression.}
\usage{
hyperparameter_regression_on_ABC(
  ABC,
  input_independent_parameters,
  formula = scale ~ pi,
  regression_method = "rf",
  num_trees = 10000,
  num_threads = NULL,
  parameter_transforms = list(pi = function(pi) log10(1 - pi), scale = function(scale)
    log10(scale)),
  parameter_back_transforms = list(pi = function(pi) 1 - 10^pi, scale = function(scale)
    10^scale),
  acceptance_threshold = 0.005,
  dist_var = "ks",
  hold_percent = 0.25,
  quantiles = seq(0 + 0.001, 1 - 0.001, by = 0.001),
  independent_quantiles = seq(0 + 0.001, 1 - 0.001, by = 0.001)
)
}
\description{
Esitmates scale based on ABC results using gam smoothing or regression forest.
}
