% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs.R
\name{logistic_growth}
\alias{logistic_growth}
\title{Calculate population size after logistic growth}
\usage{
logistic_growth(n, K, r, ...)
}
\arguments{
\item{n}{numeric. Starting population size.}

\item{K}{numeric. Carrying capacity.}

\item{r}{numeric. Logistic growth rate.}
}
\description{
Given a starting population size, find the estiamted size after one generation of
logistic growth.
}
