% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{make_G}
\alias{make_G}
\title{Make a G matrix}
\usage{
make_G(ind.genos, maf = 0.05, par = 1)
}
\arguments{
\item{ind.genos}{Unphased genotypes, rows are \emph{individuals}.}
}
\description{
Make a G matrix using the Yang et al (2010) method from genotypes.
}
