% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{make_yang_G_FBM}
\alias{make_yang_G_FBM}
\title{Yang method G matrix from a SNP big matrix}
\usage{
make_yang_G_FBM(SNPmatrix, maf = 0.05, par = 1)
}
\arguments{
\item{SNPmatrix}{File-backed matrix (FBM), where each row is an \emph{individual} and each
column is a \emph{loci}. Values are the intergers 0, 1, or 2, where 0 and 2 are homozygotes and 1 is
a heterozygote.}

\item{maf}{numeric, default 0.05. Loci with minor allele frequencies below this will be removed.}

\item{par}{numeric, default 1. Number of cores to use for calculations.}
}
\description{
Make a G matrix using the method of Yang et al (2010) from a bigstatsr
FBM (File-Backed Matrix). Uses code adapted from the AGH matrix package.
Slower than simple vectorized code or the AGHmatrix function, but runs
on huge genotype files in parallel without using huge amounts of memory.
}
