% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_description.R
\name{mark_windows}
\alias{mark_windows}
\title{Mark windows for snps.}
\usage{
mark_windows(x, sigma, chr = "chr")
}
\arguments{
\item{x}{data.frame. Must contain a "position" column and a chromosome info column.}

\item{sigma}{numeric. Size of windows, in kb.}

\item{chr}{character, default "chr". Name of chromosome info column in x.}
}
\description{
Determine which unique genomic window each snp belongs to.
}
