% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\docType{data}
\name{names_descriptive_stats}
\alias{names_descriptive_stats}
\title{Names of the descriptive statistics calculated from GWAS results.}
\format{
A character vector of length 729.
}
\usage{
names_descriptive_stats
}
\description{
Names of the descriptive statistics calculated from GWAS results.
}
\keyword{datasets}
