% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\docType{data}
\name{names_diff_stats}
\alias{names_diff_stats}
\title{Names of the statistics calculated during distribution comparisons}
\format{
A character vector of length 32.
}
\usage{
names_diff_stats
}
\description{
Names of the statistics calculated during distribution comparisons
}
\keyword{datasets}
