% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs.R
\name{normal_survival}
\alias{normal_survival}
\title{Calculate survivability probabilities using a scaled normal distribution.}
\usage{
normal_survival(phenotypes, opt_pheno, var, max.survival = 0.5, ...)
}
\arguments{
\item{phenotypes}{numeric. Vector of phenotypes.}

\item{opt_pheno}{numeric. Optimum phenotype.}

\item{var}{numeric. Variance of surivial distribution.}

\item{max.survival}{numeric. Maximum possible survival probability.}
}
\description{
For a given optimum phenotype and surivival variance, calculate survival probabilities from a
normal distribution.
}
\details{
Surival probabilities are given by \code{\link[stats]{pnorm}}. For values above the optimum phenotype,
the surivial is given by 1 - the resulting value. If a maxium survival other than 0.5 is given,
the surivivals are scaled such that individuals with the optimum phenotype will have the provided
survival probability.
}
