% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{number_descriptive_stats}
\alias{number_descriptive_stats}
\title{Number of descriptive parameters produced.}
\format{An object of class \code{numeric} of length 1.}
\usage{
number_descriptive_stats
}
\description{
Number of descriptive parameters produced.
}
\references{
Hemstrom et al (in prep)
}
\keyword{datasets}
