% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs.R
\name{optimum_logistic_slide}
\alias{optimum_logistic_slide}
\title{Increases the selection optimum according to logistic growth.}
\usage{
optimum_logistic_slide(opt, init_opt, scale, r, ...)
}
\arguments{
\item{opt}{numeric. Current optimum phenotype.}

\item{init_opt}{numeric. Initial optimum phenotype}

\item{scale}{numeric. The proportion of the initial optimum phenotype at which the slide will max out.}

\item{r}{numeric. The growth rate of the optimum phenotype.}
}
\description{
Increases the selection optimum according to logistic growth.
}
