% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_estimation_tools.R
\name{pred_gwas_FBM}
\alias{pred_gwas_FBM}
\title{Run a GMMAT GWAS on genotype/phenotype data.}
\usage{
pred_gwas_FBM(
  x = NULL,
  phenotypes,
  maf = 0.05,
  pass_G = NULL,
  GMMAT_infile = NULL,
  phased = F,
  par = 1,
  center = T
)
}
\arguments{
\item{x}{genotype data, default NULL. Genotype
with individuals in columns (either 1 or 2 columns per individual for unphased or phased data). Can either
be coercable to a matrix or be a \code{\link[bigstatsr]{FBM}}.}

\item{phenotypes}{numeric. Phenotype data, one numeric value per individual, in the same order as the genotype data.}

\item{maf}{numeric, default 0.05. Minor allele frequency filter to be applied to both G-matrix and GWAS. Ignored if x is NULL.}

\item{pass_G}{numeric matrix, default NULL. An \emph{n x n} genetic relatedness matrix. If NULL, calculated from x via
\code{\link[AGHmatrix]{Gmatrix}} using the Yang et al 2010 method.}

\item{GMMAT_infile}{character, default NULL. If provided, the file path to an infile for \code{\link[GMMAT]{glmm.score}}. If NULL,
will be written from x.}

\item{phased}{logical, default F. Is the data in x phased (in two columns per individual)? Ignored if x is NULL.}

\item{par}{numeric, default 1. Number of parallel cores to use for G matrix creation. Ignored if x is NULL.}

\item{center}{logical, default T. If TRUE, the phenotypes will be centered (set to mean 0) prior to GWAS calculation}
}
\description{
Run a GMMAT GWAS on genotype/phenotype data.
}
