% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{rbayesB}
\alias{rbayesB}
\title{Get random draws from the distribution used for bayesB regressions.}
\usage{
rbayesB(n, pi, d.f, scale)
}
\arguments{
\item{n}{numeric. Number of draws to make from the distribution}

\item{pi}{numeric. Probability that any one site has zero effect}

\item{d.f}{numeric. Degrees of freedom for the scaled-inverse chi-squared
distribution}

\item{scale}{numeric. Scale/shape parameter for the scaled-inverse
chi-squared distribution.}
}
\description{
Generate any number of random values drawn from the distribution used for
bayesB genomic regressions.
}
\details{
Under a bayesB model, the effect of each site is drawn from a distribution
where var(g) == 0 with prob \emph{pi}, and is otherwise drawn from a scaled
t distribution with degrees of freedom \emph{d.f} and scale
\emph{scale}.
}
