% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{reasonable_transform}
\alias{reasonable_transform}
\title{Reasonable parameter transformations}
\usage{
reasonable_transform(parameters = c("pi", "scale"))
}
\arguments{
\item{parameters}{character, default c("pi", "scale"). Parameters for which to fetch reasonable transformation functions.}
}
\value{
A nested, named list containing forward and back transforms for the named parameters.
}
\description{

}
\examples{
# for pi
reasonable_transform("pi")

# for pi and scale
reasonable_transform(c("pi", "scale"))

# when there aren't any banked transformations, returns the identity function
reasonable_transform(c("pi", "not_a_parameter", "scale"))
}
