% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs.R
\name{sample_joint_quantile}
\alias{sample_joint_quantile}
\title{Sample hyperparameter values from hyperparameter regression/random forest results.}
\usage{
sample_joint_quantile(n, reg)
}
\arguments{
\item{n}{numeric. The number of samples to draw.}

\item{reg}{regression results from \code{\link{hyperparameter_regression_on_ABC}}.}
}
\value{
A data.frame containing the sampled hyperparameter values.
}
\description{
Samples hyperparameters from the joint quantile distribution resulting from a regression on ABC results
and a random forest.
}
\author{
William Hemstrom
}
