% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{sprinkle_missing_data}
\alias{sprinkle_missing_data}
\title{Adds missing data to a genotype dataset.}
\usage{
sprinkle_missing_data(
  x,
  missing_dist = function(x) {     missing_dist <- rnorm(x, mean = 0.05, sd = 0.02)    
    missing_dist[missing_dist < 0] <- 0     return(missing_dist) },
  missing_code = -9
)
}
\arguments{
\item{x}{object coercable to numeric matrix containing genotype calls. Rows are loci, columns are samples. Each sample
should be the genotype for a gene copy}
}
\description{
Adds missing data to a genotype dataset.
}
